Many Changes
Game is in a working state with a beginner quest, room to sleep in,
loading screen, a destroyed city with a "task" of finding what happened.
Also contains a new NPC who will be used for the plot. "Robot Kyle"
Right now you can interact with him but its not meaningful text.